BE/laptop_store/requirements.txt
django>=4.2,<4.3
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
mysqlclient>=2.2,<3.0
bcrypt>=4.1,<5.0
python-dotenv>=1.0,<2.0
Pillow>=10.0
scikit-learn>=1.4,<2.0
numpy>=1.26
celery>=5.3,<6.0
redis>=5.0,<6.0
google-genai>=1.0,<2.0
